Cover SiteMapExtractorBase
